Move unit tests for utilities
