+++ platform-sdk/consensus-concurrent/src/main/java/module-info.java
@@ -9,7 +9,7 @@ module org.hiero.consensus.concurrent {
     exports org.hiero.consensus.concurrent.framework.internal;
     exports org.hiero.consensus.concurrent.manager;
     exports org.hiero.consensus.concurrent.pool;
-    exports org.hiero.consensus.concurrent.utility.throttle;
+    exports org.hiero.consensus.concurrent.throttle;
 
     requires transitive com.swirlds.base;
     requires transitive com.swirlds.config.api;
