+++ platform-sdk/consensus-concurrent/src/test/java/org/hiero/consensus/concurrent/throttle/RateLimitedLoggerTests.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.utility.throttle;
+package org.hiero.consensus.concurrent.throttle;
 
 import static com.swirlds.logging.legacy.LogMarker.EXCEPTION;
 import static org.hiero.base.utility.test.fixtures.RandomUtils.getRandomPrintSeed;
@@ -18,7 +18,6 @@ import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.Marker;
 import org.hiero.base.CompareTo;
-import org.hiero.consensus.concurrent.utility.throttle.RateLimitedLogger;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
