+++ platform-sdk/consensus-concurrent/src/test/java/org/hiero/consensus/concurrent/throttle/RateLimiterTests.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.utility.throttle;
+package org.hiero.consensus.concurrent.throttle;
 
 import static org.hiero.base.CompareTo.isLessThan;
 import static org.hiero.base.utility.test.fixtures.RandomUtils.getRandomPrintSeed;
@@ -9,7 +9,6 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import com.swirlds.base.test.fixtures.time.FakeTime;
 import java.time.Duration;
 import java.util.Random;
-import org.hiero.consensus.concurrent.utility.throttle.RateLimiter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
