+++ platform-sdk/consensus-event-intake-concurrent/src/main/java/org/hiero/consensus/event/intake/concurrent/ConcurrentEventIntakeProcessor.java
@@ -24,7 +24,7 @@ import java.util.function.Function;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hiero.base.crypto.BytesSignatureVerifier;
-import org.hiero.consensus.concurrent.utility.throttle.RateLimitedLogger;
+import org.hiero.consensus.concurrent.throttle.RateLimitedLogger;
 import org.hiero.consensus.crypto.EventHasher;
 import org.hiero.consensus.event.IntakeEventCounter;
 import org.hiero.consensus.event.validation.EventFieldValidator;
