+++ platform-sdk/consensus-gossip-impl/src/test/java/org/hiero/consensus/gossip/impl/network/NetworkUtilsTest.java
@@ -4,7 +4,7 @@ package org.hiero.consensus.gossip.impl.network;
 import com.swirlds.base.time.Time;
 import java.time.Duration;
 import javax.net.ssl.SSLException;
-import org.hiero.consensus.concurrent.utility.throttle.RateLimiter;
+import org.hiero.consensus.concurrent.throttle.RateLimiter;
 import org.hiero.consensus.gossip.impl.test.fixtures.sync.FakeConnection;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
