+++ platform-sdk/consensus-metrics/src/main/java/org/hiero/consensus/metrics/extensions/StandardFractionalTimer.java
@@ -14,7 +14,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hiero.base.StackTrace;
 import org.hiero.base.utility.ByteUtils;
-import org.hiero.consensus.concurrent.utility.throttle.RateLimitedLogger;
+import org.hiero.consensus.concurrent.throttle.RateLimitedLogger;
 import org.hiero.consensus.metrics.FunctionGauge;
 
 /**
