+++ platform-sdk/swirlds-component-framework/src/main/java/com/swirlds/component/framework/model/internal/monitor/HealthMonitorLogger.java
@@ -14,7 +14,7 @@ import java.util.Map;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hiero.base.CompareTo;
-import org.hiero.consensus.concurrent.utility.throttle.RateLimitedLogger;
+import org.hiero.consensus.concurrent.throttle.RateLimitedLogger;
 
 /**
  * Encapsulates logging for the wiring health monitor.
